requirements.txt
openai
googlesearch-python
requests
bs4